+++ src/main/kotlin/com/nylas/resources/Auth.kt
@@ -136,7 +136,7 @@ class Auth(private val client: NylasClient) {
    */
   @Throws(NylasApiError::class, NylasSdkTimeoutError::class)
   fun detectProvider(params: ProviderDetectParams): Response<ProviderDetectResponse> {
-    val path = "v3/grants/connect/providers/detect"
+    val path = "v3/providers/detect"
     val responseType = Types.newParameterizedType(Response::class.java, ProviderDetectResponse::class.java)
 
     return client.executePost(path, responseType, queryParams = params)
@@ -165,7 +165,11 @@ class Auth(private val client: NylasClient) {
       .adapter(UrlForAuthenticationConfig::class.java)
       .toJson(config)
     JsonHelper.jsonToMap(json).forEach { (key, value) ->
-      url.addQueryParameter(key, value.toString())
+      if(key == "scope") {
+        url.addQueryParameter("scope", config.scope?.joinToString(separator = " "))
+      } else {
+        url.addQueryParameter(key, value.toString())
+      }
     }
 
     return url